@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class Genre < ActiveRecord::Base
+class Genre < ApplicationRecord
   has_many :albums
 
   validates :name, presence: true, uniqueness: {case_sensitive: false}
