@@ -22,7 +22,7 @@ Rails.application.configure do
 
   # Disable serving static files from the `/public` folder by default since
   # Apache or NGINX already handles this.
-  config.serve_static_files = ENV['RAILS_SERVE_STATIC_FILES'].present?
+  config.public_file_server.enabled = ENV['RAILS_SERVE_STATIC_FILES'].present?
 
   # Compress JavaScripts and CSS.
   config.assets.js_compressor = :uglifier
@@ -97,4 +97,9 @@ Rails.application.configure do
     enable_starttls_auto: true
   }
   config.action_mailer.default_url_options = {host: host}
+
+  # Simple Rails log rotation. Up to six 100MB-sized log chunks, 1 current
+  # and 5 old will be kept.
+  config.logger = ActiveSupport::Logger.new(config.paths["log"].first,
+                                            5, 100 * 1024 * 1024)
 end
